Stylecop changes, comments.
