+++ src/AutoStep/Compiler/LinkResult.cs
@@ -14,6 +14,7 @@ namespace AutoStep.Compiler
         /// </summary>
         /// <param name="success">Indicates a successful link.</param>
         /// <param name="messages">The set of messages resulting from a link.</param>
+        /// <param name="referencedSources">The set of all referenced step definition sources found by the linker.</param>
         /// <param name="output">The built output (if the link succeeded).</param>
         public LinkResult(bool success, IEnumerable<CompilerMessage> messages, IEnumerable<IStepDefinitionSource>? referencedSources = null, BuiltFile? output = null)
             : base(success, messages, output)
@@ -26,8 +27,14 @@ namespace AutoStep.Compiler
             ReferencedSources = referencedSources ?? Enumerable.Empty<IStepDefinitionSource>();
         }
 
-        public bool AnyIssues { get; internal set; }
+        /// <summary>
+        /// Gets a value indicating whether any Errors or Warnings were found.
+        /// </summary>
+        public bool AnyIssues { get; }
 
+        /// <summary>
+        /// Gets the set of all referenced step definition sources found by the linker.
+        /// </summary>
         public IEnumerable<IStepDefinitionSource> ReferencedSources { get; }
     }
 }
