+++ src/AutoStep/Definitions/FileStepDefinitionSource.cs
@@ -1,60 +1,56 @@
 ﻿using System;
 using System.Collections.Generic;
 using System.Linq;
-using System.Text;
-using AutoStep.Compiler;
-using AutoStep.Elements;
 
 namespace AutoStep.Definitions
 {
-    public interface IUpdatableStepDefinitionSource : IStepDefinitionSource
-    {
-        DateTime GetLastModifyTime();
-    }
-
-    internal class FileStepDefinition : StepDefinition
-    {
-        public FileStepDefinition(IStepDefinitionSource source, StepDefinitionElement element)
-            : base(
-                  source,
-                  element?.Type ?? throw new ArgumentNullException(nameof(element)),
-                  element.Declaration!) // The declaration is validated by FileStepDefinitionSource before instantiating.
-        {
-            Definition = element;
-        }
-
-        public override bool IsSameDefinition(StepDefinition def)
-        {
-            return Type == def.Type && def.Declaration == def.Declaration;
-        }
-    }
-
+    /// <summary>
+    /// Defines a step definition source backed by a project file that has its own step definitions.
+    /// </summary>
     internal class FileStepDefinitionSource : IUpdatableStepDefinitionSource
     {
         /// <summary>
         /// Initializes a new instance of the <see cref="FileStepDefinitionSource"/> class.
         /// </summary>
-        /// <param name="file"></param>
+        /// <param name="file">The file to get steps from.</param>
         public FileStepDefinitionSource(ProjectFile file)
         {
             File = file;
         }
 
+        /// <summary>
+        /// Gets the unique ID of the source (the file).
+        /// </summary>
         public string Uid => File.Path;
 
+        /// <summary>
+        /// Gets the name of the file.
+        /// </summary>
         public string Name => File.Path;
 
+        /// <summary>
+        /// Gets the backing project file.
+        /// </summary>
         public ProjectFile File { get; }
 
+        /// <summary>
+        /// Gets the last modification time of the source, which is equal to the last compilation
+        /// time of the file.
+        /// </summary>
+        /// <returns>The UTC stamp for the file modification time.</returns>
         public DateTime GetLastModifyTime()
         {
             // The last modify time of the source is the last compile output, not the actual file!
             return File.LastCompileTime;
         }
 
+        /// <summary>
+        /// Gets the step definitions in the backing file's last compilation.
+        /// </summary>
+        /// <returns>The set of available definitions.</returns>
         public IEnumerable<StepDefinition> GetStepDefinitions()
         {
-            return File.LastCompileResult?.Output?.StepDefinitions.Select(d => new FileStepDefinition(this, d)) 
+            return File.LastCompileResult?.Output?.StepDefinitions.Select(d => new FileStepDefinition(this, d))
                    ?? Enumerable.Empty<FileStepDefinition>();
         }
     }
