+++ src/AutoStep/Definitions/IUpdatableStepDefinitionSource.cs
@@ -1,0 +1,9 @@
+﻿using System;
+
+namespace AutoStep.Definitions
+{
+    public interface IUpdatableStepDefinitionSource : IStepDefinitionSource
+    {
+        DateTime GetLastModifyTime();
+    }
+}
