+++ src/AutoStep/Projects/ProjectConfiguration.cs
@@ -1,0 +1,19 @@
+﻿using System;
+using System.Collections.Generic;
+using System.Text;
+using System.Threading;
+using System.Threading.Tasks;
+using AutoStep.Compiler;
+
+namespace AutoStep
+{
+
+    /// <summary>
+    /// Represents the configuration for the project.
+    /// This + environment variables = run configuration.
+    /// </summary>
+    public class ProjectConfiguration
+    {
+
+    }
+}
