--- src/AutoStep/Projects/ProjectFilter.cs
@@ -1,17 +1,0 @@
-﻿using System;
-using System.Collections.Generic;
-using System.Text;
-using System.Threading.Tasks;
-using AutoStep.Compiler;
-using Microsoft.Extensions.DependencyInjection;
-
-namespace AutoStep.Execution
-{
-    /// <summary>
-    /// Defines the filter used to specify what tests to run.
-    /// </summary>
-    public class ProjectFilter
-    {
-
-    }
-}
