+++ tests/AutoStep.Tests/Projects/ProjectFileContentSourceTests.cs
@@ -11,7 +11,7 @@ namespace AutoStep.Tests.Projects
         [Fact]
         public void ConstructorNullFileProviderThrowsException()
         {
-            Action act = () => new ProjectFileContentSource("/path", null);
+            Action act = () => new FileProviderContentSource("/path", null);
             act.Should().Throw<ArgumentNullException>();
         }
 
@@ -20,7 +20,7 @@ namespace AutoStep.Tests.Projects
         {
             var fileProvider = new DummyFileProvider("content", DateTime.Now);
 
-            var contentSource = new ProjectFileContentSource("/path", fileProvider);
+            var contentSource = new FileProviderContentSource("/path", fileProvider);
 
             contentSource.GetLastContentModifyTime().Kind.Should().Be(DateTimeKind.Utc);
         }
@@ -30,7 +30,7 @@ namespace AutoStep.Tests.Projects
         {
             var fileProvider = new DummyFileProvider("content", DateTime.Now);
 
-            var contentSource = new ProjectFileContentSource("/path", fileProvider);
+            var contentSource = new FileProviderContentSource("/path", fileProvider);
 
             contentSource.GetContentAsync().GetAwaiter().GetResult().Should().Be("content");
         }
